[SQL] Break up RustFileWriter into several classes
